requirements.txt
Flask
Pillow
inference-sdk==0.9.17
gunicorn
requests

